Image decoration (#2)
* Change dev script to allow mobile access in network

* Wrap keybindings in a compartment

* Add tab stops outside link completions

* Add image decorations

* Make image optional and enable custom styling

* Update dependencies

* Make the liquid parser more resiliant

v1 of the liquid parsing didn't recognize control flow blocks (e.g. {%  %})
which gets rid of a lot of powerful features of the editor. This leverages
regex a little more intensely to parse the resulting strings and ensures that
we have more flexible and readable control over the resulting code we find

* Fix the image-decoration and keybinding